+++ src/components/feature/ToggleProvider.tsx
@@ -1,12 +1,13 @@
 'use client';
 
 import { createContextWithHook } from '@/utils/createContextWithHook';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useId, useRef, useState } from 'react';
 import { PassProps } from '../util/PassProps';
 
 type ToggleContextProps = {
   state: boolean;
   toggleState: (newState?: boolean) => void;
+  id: string;
 };
 
 const [ToggleContext, useToggleContext] =
@@ -19,7 +20,7 @@ type ToggleProviderProps = React.PropsWithChildren & {
 
 export function ToggleProvider({ children, onChange, initialState = false }: ToggleProviderProps) {
   const [state, setState] = useState(initialState);
-
+  const id = useId();
   const toggleState = (newState?: boolean) => {
     const fn = setState;
     fn(newState !== undefined ? newState : !state);
@@ -27,14 +28,20 @@ export function ToggleProvider({ children, onChange, initialState = false }: Tog
 
   useEffect(() => toggleState(initialState), [initialState]);
   useEffect(() => onChange && onChange(state), [state]);
-  return <ToggleContext.Provider value={{ state, toggleState }}>{children}</ToggleContext.Provider>;
+  return (
+    <ToggleContext.Provider value={{ state, toggleState, id: id.split(':').at(1) }}>
+      {children}
+    </ToggleContext.Provider>
+  );
 }
 
+const triggerClassName = 'toggle-provider-trigger';
 ToggleProvider.Trigger = function ({ children, ...props }) {
-  const { toggleState } = useToggleContext();
+  const { toggleState, id } = useToggleContext();
   return (
     <PassProps
       {...props}
+      className={`${triggerClassName}-${id}`}
       onClick={() => toggleState()}>
       {children}
     </PassProps>
@@ -42,12 +49,21 @@ ToggleProvider.Trigger = function ({ children, ...props }) {
 };
 
 ToggleProvider.Target = function ({ children, hideOnClickOutside = false }) {
-  const { state, toggleState } = useToggleContext();
+  const { state, toggleState, id: toggleProviderId } = useToggleContext();
   const ref = useRef<HTMLElement | null>(null);
 
+  //Closes the target if clicking outside of it, and not clicking on any of its triggers.
   const handleClickOutside = e => {
-    if (hideOnClickOutside && ref.current && !ref.current.contains(e.target)) {
-      console.log('Calling handleClickOutside');
+    const safeId = toggleProviderId;
+    const closestTrigger = e.target.closest(`.${triggerClassName}-${safeId}`);
+    if (
+      hideOnClickOutside &&
+      ref.current &&
+      //The click happens outside the target
+      !ref.current.contains(e.target) &&
+      //The click does not happen on any of the triggers
+      closestTrigger === null
+    ) {
       toggleState(false);
     }
   };